Another tweak to qemu-setup.sh

+++ sources/native/setup.sh
@@ -4,11 +4,12 @@
 
 # Create some temporary directories at the root level
 mkdir -p /{proc,sys,etc}
-ln -s /tools/bin /bin
-ln -s /tools/lib /lib
+[ ! -e /bin ] && ln -s /tools/bin /bin
+[ ! -e /lib ] && ln -s /tools/lib /lib
 
 # Populate /dev
 mount -t sysfs /sys /sys
+mount -t tmpfs /dev /dev
 mdev -s
 
 # Setup network for QEMU
@@ -16,7 +17,7 @@ mount -t proc /proc /proc
 echo "nameserver 10.0.2.3" > /etc/resolv.conf
 ifconfig eth0 10.0.2.15
 route add default gw 10.0.2.2
-rdate time-b.nist.gov
+#rdate time-b.nist.gov
 
 exec /tools/bin/busybox ash
